Some changes.
For one, I realized I messed up the ordering of the columns.  I have not
heavily tested this change.  :(

For the rest, just trying to reword.  I think I'm going to stick a fork
in this and move on to another problem.  I'm sure I'll come back to this
later.  May just redo it.