Finally started playing with a cube.
I've been toying with this idea for a long time now.  I'd prefer a 3d
styling of the cube, I think.  I confess this styling is somewhat
mesmerizing to watch animate.

My plan is sometime next week to build the logic that will search for
a solution from a given permutation.