Amusing failure at finding amusingly simple algorithm.
I go over it more in detail in the post.  Pretty sure I did not find the
algorithm.

I did finally figure out how to tweak the output of a graphviz chart.  I
think I may have to make that a tip at work.  What are the odds I can
remember how to do this otherwise? :)