I finally wrote a BDD I had bouncing in my head.
I drug my feet to write this, I think largely from fear that it
wouldn't work.  I think it is fair that this would not be a good idea
by itself on a really large list of numbers.  Maybe a ZDD would be
better?  Probably still not a good use.

Regardless, I had fun.